requirements.txt
python-telegram-bot==21.0.1
Requests==2.31.0
